refactor: move theme cubit inside core folder
